:art: update compass
